Fix - cards send to a separate component and add new budget card
