+++ src/app/dashboard/page.tsx
@@ -1,6 +1,6 @@
 import { auth } from "@/auth";
 import { Metadata } from "next";
-import { Card } from "../ui/dashboard/cards";
+import { Card, Cards } from "../ui/dashboard/cards";
 import {
   fetchCardData,
   fetchEssentialSpendDataByMonth,
@@ -14,7 +14,6 @@ import ExpensesMonthChart from "../ui/dashboard/expenses-month-chart";
 import ExpensesCategoryChart from "../ui/dashboard/expenses-category-chart";
 import Toggle from "../ui/dashboard/Toggle";
 import EssentialExpensesMonthChart from "../ui/dashboard/essential-expenses-chart";
-import dashboardMessages from "../lib/data/messages/dashboard.json";
 
 export const metadata: Metadata = {
   title: "Dashboard",
@@ -35,21 +34,6 @@ export default async function Page({
 
   const currentYear = new Date().getFullYear();
   const currentMonth = new Date().getMonth() + 1;
-
-  const {
-    totalMonthSpend,
-    totalMonthIncome,
-    totalYearSpend,
-    totalYearIncome,
-    totalMonthSpendIncome,
-    totalYearSpendIncome,
-    percentageOfIncomeSpentMonth,
-    percentageOfIncomeSpentYear,
-  } = await fetchCardData(
-    session.user.id,
-    currentMonth.toString(),
-    currentYear.toString(),
-  );
 
   const spendByMonth = await fetchSpendDataByMonth(
     session.user.id,
@@ -94,48 +78,26 @@ export default async function Page({
 
   let spendByCategoryMap;
 
-  let totalSpendValue;
-  let totalIncomeValue;
-  let totalSpendIncomeValue;
-  let totalSpendIncomePercentage;
   if (currentPeriod === "Month") {
-    totalSpendValue = totalMonthSpend;
-    totalIncomeValue = totalMonthIncome;
-    totalSpendIncomeValue = totalMonthSpendIncome;
     spendByCategoryMap = new Map(
       spendByCategoryMonthly.map((item) => [item.category, item.total]),
     );
-    const totalMonthSpendValue = Number(totalMonthSpend);
-    const totalMonthIncomeValue = Number(totalMonthIncome);
-    totalSpendIncomePercentage = percentageOfIncomeSpentMonth;
   } else {
-    totalSpendValue = totalYearSpend;
-    totalIncomeValue = totalYearIncome;
-    totalSpendIncomeValue = totalYearSpendIncome;
     spendByCategoryMap = new Map(
       spendByCategoryYearly.map((item) => [item.category, item.total]),
     );
-    const totalYearSpendValue = Number(totalYearSpend);
-    const totalYearIncomeValue = Number(totalYearIncome);
-    totalSpendIncomePercentage = percentageOfIncomeSpentYear;
   }
 
   return (
     <main>
       <h1 className={`mb-4 text-xl md:text-2xl`}>Dashboard</h1>
       <Toggle />
-      <div className="grid gap-6 sm:grid-cols-2 lg:grid-cols-4">
-        <Card title="Current Spend" value={totalSpendValue} type="month" />
-        <Card title="Current Income" value={totalIncomeValue} type="year" />
-        <Card
-          title={dashboardMessages.dashboard.cards.savings}
-          value={totalSpendIncomeValue}
-          type="spendIncome"
-        />
-        <Card
-          title={dashboardMessages.dashboard.cards.percentage}
-          value={totalSpendIncomePercentage}
-          type="percentage"
+      <div className="grid gap-6 sm:grid-cols-2 md:grid-cols-2 lg:grid-cols-4 xl:grid-cols-6">
+        <Cards
+          currentPeriod={currentPeriod}
+          currentMonth={currentMonth}
+          currentYear={currentYear}
+          userId={session.user.id}
         />
       </div>
       <div className="hidden md:block">
