+++ src/app/lib/data/messages/dashboard.json
@@ -5,8 +5,9 @@
     "cards": {
       "currentSpend": "Current Spend",
       "currentIncome": "Current Income",
-      "savings": "Savings(Income - Spend)",
-      "percentage": "Income Spent"
+      "savings": "Savings",
+      "percentage": "Income Spent",
+      "budget": "Budget"
     }
   }
 }
